addTwoNumbers.cpp
Initial commit of embarrassingly simple version of the list nodes. But for the life of me I could not figure out how to get around the fact that when you not even call just declare ListNode* solutionNodes; that it sets the first value to 0; and then when I did get it it was backwards.

Maybe some funny business in handling the notion that you can't really have a pointer to just an empty list...or can you? Feels like you should. Like yo, there's gonna be a list there.