longestSubstringWithoutRepeatingCharacters python
And the intuition that one must pass through the entire string at least once is correct. 

But kind of the bamboozle comes in how that answer value is calculated. I must find a better name for the set used for the comparison because it does not really capture what is happening.